add image_name when sync code by webhook (#2501)
* add image_name when sync code by webhook


* refactor code

servcie and webhook set container use the same code now


* remove comments


* fix import order and change comment language


---------
