+++ pkg/microservice/aslan/core/service/service/service.go
@@ -631,7 +631,7 @@ func CreateWorkloadTemplate(userName string, args *commonmodels.Service, log *za
 	}
 	// 遍历args.KubeYamls，获取 Deployment 或者 StatefulSet 里面所有containers 镜像和名称
 	args.KubeYamls = []string{args.Yaml}
-	if err := setCurrentContainerImages(args); err != nil {
+	if err := util.SetCurrentContainerImages(args); err != nil {
 		log.Errorf("Failed tosetCurrentContainerImages %s, err: %s", args.ProductName, err)
 		return err
 	}
@@ -912,7 +912,7 @@ func UpdateServiceVariables(args *commonservice.ServiceTmplObject) error {
 	currentService.RenderedYaml = util.ReplaceWrapLine(currentService.RenderedYaml)
 	currentService.KubeYamls = util.SplitYaml(currentService.RenderedYaml)
 	oldContainers := currentService.Containers
-	if err := setCurrentContainerImages(currentService); err != nil {
+	if err := util.SetCurrentContainerImages(currentService); err != nil {
 		log.Errorf("failed to ser set container images, err: %s", err)
 		//return err
 	} else if containersChanged(oldContainers, currentService.Containers) {
@@ -1493,7 +1493,7 @@ func ensureServiceTmpl(userName string, args *commonmodels.Service, log *zap.Sug
 
 		// since service may contain go-template grammar, errors may occur when parsing as k8s workloads
 		// errors will only be logged here
-		if err := setCurrentContainerImages(args); err != nil {
+		if err := util.SetCurrentContainerImages(args); err != nil {
 			log.Errorf("failed to ser set container images, err: %s", err)
 			//return err
 		}
@@ -1531,131 +1531,6 @@ func distinctEnvServices(productName string) (map[string][]*commonmodels.Product
 		}
 	}
 	return serviceMap, nil
-}
-
-func setCurrentContainerImages(args *commonmodels.Service) error {
-	var srvContainers []*commonmodels.Container
-	for _, data := range args.KubeYamls {
-		yamlDataArray := util.SplitYaml(data)
-		for index, yamlData := range yamlDataArray {
-			resKind := new(types.KubeResourceKind)
-			//在Unmarshal之前填充渲染变量{{.}}
-			yamlData = config.RenderTemplateAlias.ReplaceAllLiteralString(yamlData, "ssssssss")
-			// replace $Service$ with service name
-			yamlData = config.ServiceNameAlias.ReplaceAllLiteralString(yamlData, args.ServiceName)
-			// replace $Product$ with product name
-			yamlData = config.ProductNameAlias.ReplaceAllLiteralString(yamlData, args.ProductName)
-
-			if err := yaml.Unmarshal([]byte(yamlData), &resKind); err != nil {
-				return fmt.Errorf("unmarshal ResourceKind error: %v", err)
-			}
-
-			if resKind == nil {
-				if index == 0 {
-					continue
-				}
-				return errors.New("nil Resource Kind")
-			}
-
-			if resKind.Kind == setting.Deployment || resKind.Kind == setting.StatefulSet || resKind.Kind == setting.Job {
-				containers, err := getContainers(yamlData)
-				if err != nil {
-					return fmt.Errorf("GetContainers error: %v", err)
-				}
-
-				srvContainers = append(srvContainers, containers...)
-			} else if resKind.Kind == setting.CronJob {
-				containers, err := getCronJobContainers(yamlData)
-				if err != nil {
-					return fmt.Errorf("GetCronjobContainers error: %v", err)
-				}
-				srvContainers = append(srvContainers, containers...)
-			}
-		}
-	}
-
-	args.Containers = uniqueSlice(srvContainers)
-	return nil
-}
-
-func getContainers(data string) ([]*commonmodels.Container, error) {
-	containers := make([]*commonmodels.Container, 0)
-
-	res := new(types.KubeResource)
-	if err := yaml.Unmarshal([]byte(data), &res); err != nil {
-		return containers, fmt.Errorf("unmarshal yaml data error: %v", err)
-	}
-
-	for _, val := range res.Spec.Template.Spec.Containers {
-
-		if _, ok := val["name"]; !ok {
-			return containers, errors.New("yaml file missing name key")
-		}
-
-		nameStr, ok := val["name"].(string)
-		if !ok {
-			return containers, errors.New("error name value")
-		}
-
-		if _, ok := val["image"]; !ok {
-			return containers, errors.New("yaml file missing image key")
-		}
-
-		imageStr, ok := val["image"].(string)
-		if !ok {
-			return containers, errors.New("error image value")
-		}
-
-		container := &commonmodels.Container{
-			Name:      nameStr,
-			Image:     imageStr,
-			ImageName: util.ExtractImageName(imageStr),
-		}
-
-		containers = append(containers, container)
-	}
-
-	return containers, nil
-}
-
-func getCronJobContainers(data string) ([]*commonmodels.Container, error) {
-	containers := make([]*commonmodels.Container, 0)
-
-	res := new(types.CronjobResource)
-	if err := yaml.Unmarshal([]byte(data), &res); err != nil {
-		return containers, fmt.Errorf("unmarshal yaml data error: %v", err)
-	}
-
-	for _, val := range res.Spec.Template.Spec.Template.Spec.Containers {
-
-		if _, ok := val["name"]; !ok {
-			return containers, errors.New("yaml file missing name key")
-		}
-
-		nameStr, ok := val["name"].(string)
-		if !ok {
-			return containers, errors.New("error name value")
-		}
-
-		if _, ok := val["image"]; !ok {
-			return containers, errors.New("yaml file missing image key")
-		}
-
-		imageStr, ok := val["image"].(string)
-		if !ok {
-			return containers, errors.New("error image value")
-		}
-
-		container := &commonmodels.Container{
-			Name:      nameStr,
-			Image:     imageStr,
-			ImageName: util.ExtractImageName(imageStr),
-		}
-
-		containers = append(containers, container)
-	}
-
-	return containers, nil
 }
 
 func updateGerritWebhookByService(lastService, currentService *commonmodels.Service) error {
@@ -1750,18 +1625,4 @@ func syncGerritLatestCommit(service *commonmodels.Service) (*systemconfig.CodeHo
 		Message: commit.Message,
 	}
 	return ch, nil
-}
-
-func uniqueSlice(elements []*commonmodels.Container) []*commonmodels.Container {
-	existing := make(map[string]bool)
-	ret := make([]*commonmodels.Container, 0)
-	for _, elem := range elements {
-		key := elem.Name
-		if _, ok := existing[key]; !ok {
-			ret = append(ret, elem)
-			existing[key] = true
-		}
-	}
-
-	return ret
 }
