+++ pkg/microservice/aslan/core/workflow/service/webhook/gerrit.go
@@ -300,7 +300,7 @@ func ensureServiceTmpl(userName string, args *commonmodels.Service, log *zap.Sug
 		}
 
 		// 遍历args.KubeYamls，获取 Deployment 或者 StatefulSet 里面所有containers 镜像和名称
-		if err := setCurrentContainerImages(args); err != nil {
+		if err := util.SetCurrentContainerImages(args); err != nil {
 			return err
 		}
